Move tokenize to own file
